Convert camera to geographic coordinates and bridge Double matrices to Float for shaders
Camera now uses CLLocationCoordinate2D + zoom/bearing/pitch instead of a
generic Cartesian position/target. Matrices compute in double4x4 on the
CPU for precision at geographic scales (~1m wobble per Float scalar at
Earth radius), then downcast to simd_float4x4 at the renderer boundary
via a new SIMD+Conversion extension. This is the standard CPU→GPU
precision pattern for map renderers — bearing and pitch are stored on
the camera but not yet applied; that's Phase 2.

+++ Mir/Sources/Mir/Views/Map/MapMetalKitView+Coordinator.swift
@@ -66,7 +66,7 @@ extension MapMetalKitView {
         // MARK: - MTKViewDelegate
 
         func mtkView(_ view: MTKView, drawableSizeWillChange size: CGSize) {
-            renderer?.scene.camera.aspectRatio = Float(size.width / size.height)
+            renderer?.scene.camera.aspectRatio = size.width / size.height
         }
 
         /// A delegate method called for each frame to be rendered.
